+++ rector.php
@@ -4,14 +4,16 @@ declare(strict_types=1);
 
 use Rector\CodeQuality\Rector\Class_\InlineConstructorDefaultToPropertyRector;
 use Rector\Config\RectorConfig;
-use Rector\Php56\Rector\FunctionLike\AddDefaultValueForUndefinedVariableRector;
 use Rector\Php74\Rector\Closure\ClosureToArrowFunctionRector;
 use Rector\Set\ValueObject\LevelSetList;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->paths([
         __DIR__ . '/src',
-        __DIR__ . '/tests',
+        /**
+         * Disabled ./tests directory due to different branches with main package when testing
+         */
+        // __DIR__ . '/tests',
     ]);
 
     // register a single rule
@@ -24,6 +26,5 @@ return static function (RectorConfig $rectorConfig): void {
 
     $rectorConfig->skip([
         ClosureToArrowFunctionRector::class,
-        AddDefaultValueForUndefinedVariableRector::class,
     ]);
 };