+++ src/DQLQueryBuilder.php
@@ -5,9 +5,12 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Pgsql;
 
 use Yiisoft\Db\Expression\ArrayExpression;
+use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\JsonExpression;
 use Yiisoft\Db\Pgsql\Builder\ArrayExpressionBuilder;
+use Yiisoft\Db\Pgsql\Builder\StructuredExpressionBuilder;
+use Yiisoft\Db\Pgsql\Builder\ExpressionBuilder;
 use Yiisoft\Db\Pgsql\Builder\JsonExpressionBuilder;
 use Yiisoft\Db\QueryBuilder\AbstractDQLQueryBuilder;
 use Yiisoft\Db\QueryBuilder\Condition\LikeCondition;
@@ -50,6 +53,8 @@ final class DQLQueryBuilder extends AbstractDQLQueryBuilder
         return array_merge(parent::defaultExpressionBuilders(), [
             ArrayExpression::class => ArrayExpressionBuilder::class,
             JsonExpression::class => JsonExpressionBuilder::class,
+            StructuredExpression::class => StructuredExpressionBuilder::class,
+            Expression::class => ExpressionBuilder::class,
         ]);
     }
 }